% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calCloneLLR}
\alias{calCloneLLR}
\title{calculate the log likelihood ratio of each clonotype, accounting the uncertainty of cluster label assignment}
\usage{
calCloneLLR(cellData, prob.mat, cloneID.x = NULL, verbose = F)
}
\arguments{
\item{cellData}{data.table. Each line for a cell, and these columns as required: `majorCluster`}

\item{prob.mat}{matrix Lines for cells, and columns for assignment probability of the majorClusters. row names and column names are required}

\item{cloneID.x}{character. If specified, it will subset the cellData using clonotype id  (default: NULL)}

\item{verbose}{logical. control the returned data types. (default: FALSE)}
}
\value{
a list (verbose==T) or data.table (verbose==F)
}
\description{
calculate the log likelihood ratio of each clonotype, accounting the uncertainty of cluster label assignment
}
