% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{calIndex}
\alias{calIndex}
\alias{Startrac.calIndex}
\alias{calIndex,Startrac-method}
\alias{calIndex,Startrac-method}
\alias{calIndex}
\title{Calculate cluster level indices}
\usage{
Startrac.calIndex(object)

\S4method{calIndex}{Startrac}(object)
}
\arguments{
\item{object}{A Startrac object}
}
\value{
an object of class \code{Startrac}
}
\description{
Calculate cluster level indices
}
