% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calTissueDist}
\alias{calTissueDist}
\title{calculate Startrac.dist (tissue distribution preference)}
\usage{
calTissueDist(
  dat.tb,
  byPatient = F,
  colname.cluster = "majorCluster",
  colname.patient = "patient",
  colname.tissue = "loc"
)
}
\arguments{
\item{dat.tb}{data.frame. Each line for a cell, and these columns as required: `majorCluster`, `loc`}

\item{byPatient}{logical. whether calculate the index for each patient. (default: FALSE)}

\item{colname.cluster}{character. which column specify the cluster (default: "majorCluster")}

\item{colname.patient}{character. which column specify the patient  (default: "patient")}

\item{colname.tissue}{character. which column specify the tissue  (default: "loc")}
}
\value{
an array full of R_{o/e}
}
\description{
calculate Startrac.dist (tissue distribution preference)
}
\details{
calculate Startrac.dist (tissue distribution preference) which is based on Chisquare test.
}
