% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do.table.fisher}
\alias{do.table.fisher}
\title{perform fisher test to identify the significance of each cell in the count.dist}
\usage{
do.table.fisher(count.dist, sum.col = NULL, sum.row = NULL)
}
\arguments{
\item{count.dist}{table;}

\item{sum.col}{vector;}

\item{sum.row}{vector;}
}
\description{
perform fisher test to identify the significance of each cell in the count.dist
}
