% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.cloneDat.Zhang2018}
\alias{example.cloneDat.Zhang2018}
\title{TCR clonotype data extracted from the publication by Zhang et al.}
\source{
\url{https://www.nature.com/articles/s41586-018-0694-x}

TCR clonotype data in the TXT file
}
\description{
TCR clonotype data extracted from the publication by Zhang et al.
}
