% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,Startrac-method}
\title{initialize method for Startrac}
\usage{
\S4method{initialize}{Startrac}(.Object, cell.data, aid = "AID")
}
\arguments{
\item{object}{A Startrac object}
}
\value{
an object of class \code{Startrac}
}
\description{
initialize method for Startrac
}
