% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mcol.gini_simpson}
\alias{mcol.gini_simpson}
\title{gini_simpson of each column of the input matrix}
\usage{
mcol.gini_simpson(x)
}
\arguments{
\item{x}{matrix;}
}
\description{
gini_simpson of each column of the input matrix
}
