% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{pIndex}
\alias{pIndex}
\alias{Startrac.pIndex}
\alias{pIndex,Startrac-method}
\alias{pIndex,Startrac-method}
\alias{pIndex}
\title{Calculate pairwise indices}
\usage{
Startrac.pIndex(object, cores)

\S4method{pIndex}{Startrac}(object, cores = NULL)
}
\arguments{
\item{object}{A Startrac object}

\item{cores}{number of core to be used. Passed to doParallel::registerDoParallel. default: NULL.}
}
\value{
an object of class \code{Startrac}
}
\description{
Calculate pairwise indices
}
