% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{StartracOut.plot}
\alias{plot,StartracOut-method}
\title{plot the indexes}
\usage{
StartracOut.plot(obj, index.type, byPatient)

\S4method{plot}{StartracOut}(obj, index.type = "cluster.all",
  byPatient = F)
}
\arguments{
\item{obj}{A object of StartracOut}

\item{index.type}{one of "cluster.all", "pairwise.migr", "pairwise.tran". (default:"cluster.all")}

\item{byPatient}{logical. plot indexes of each patient (default: FALSE)}
}
\value{
a ggplot2 object or Heatmap-class object
}
\description{
plot the indexes
}
